+++ addon/components/plugins/indentation/indentation-menu.ts
@@ -4,9 +4,11 @@ import { liftListItem, sinkListItem } from 'prosemirror-schema-list';
 import SayController from '@lblod/ember-rdfa-editor/core/say-controller';
 import { chainCommands } from 'prosemirror-commands';
 import { indentNode } from '@lblod/ember-rdfa-editor/commands';
+import { NodeType } from 'prosemirror-model';
 
 type Args = {
   controller: SayController;
+  allowedTypes?: NodeType[] | NodeType;
 };
 
 export default class IndentationMenuComponent extends Component<Args> {
@@ -18,11 +20,18 @@ export default class IndentationMenuComponent extends Component<Args> {
     return this.controller.schema;
   }
 
+  get allowedTypes() {
+    if (!this.args.allowedTypes)
+      return [this.schema.nodes.paragraph, this.schema.nodes.heading];
+    if (this.args.allowedTypes instanceof Array) return this.args.allowedTypes;
+    return [this.args.allowedTypes];
+  }
+
   get indentCommand() {
     return chainCommands(
       sinkListItem(this.schema.nodes.list_item),
       indentNode({
-        types: [this.schema.nodes.paragraph, this.schema.nodes.heading],
+        types: this.allowedTypes,
         direction: 1,
         predicate: (node, pos, parent) => {
           return parent?.type !== this.schema.nodes.list_item;
@@ -35,7 +44,7 @@ export default class IndentationMenuComponent extends Component<Args> {
     return chainCommands(
       liftListItem(this.controller.schema.nodes.list_item),
       indentNode({
-        types: [this.schema.nodes.paragraph, this.schema.nodes.heading],
+        types: this.allowedTypes,
         direction: -1,
         predicate: (node, pos, parent) => {
           return parent?.type !== this.schema.nodes.list_item;