+++ addon/core/keymap.ts
@@ -14,6 +14,7 @@ import {
   exitCode,
   joinBackward,
   joinForward,
+  liftEmptyBlock,
   newlineInCode,
   selectAll,
   selectNodeBackward,
@@ -104,6 +105,8 @@ const del = chainCommands(
 /// * **Delete** and **Mod-Delete** to `deleteSelection`, `joinForward`, `selectNodeForward`
 /// * **Mod-Delete** to `deleteSelection`, `joinForward`, `selectNodeForward`
 /// * **Mod-a** to `selectAll`
+///
+/// `undo` and `redo` pc keybindings are overwritten in embedded-controller!
 export const pcBaseKeymap: Keymap = (schema, options) => ({
   'Mod-z': undo,
   'Mod-Z': undo,
@@ -151,6 +154,24 @@ export const macBaseKeymap: Keymap = (schema, options) => {
     'Alt-d': pcmap['Mod-Delete'],
     'Ctrl-a': selectTextblockStart,
     'Ctrl-e': selectTextblockEnd,
+  };
+};
+
+export const embeddedEditorBaseKeymap: Keymap = (schema) => {
+  return {
+    'Mod-b': toggleMarkAddFirst(schema.marks.strong),
+    'Mod-B': toggleMarkAddFirst(schema.marks.strong),
+    'Mod-i': toggleMarkAddFirst(schema.marks.em),
+    'Mod-I': toggleMarkAddFirst(schema.marks.em),
+    'Mod-u': toggleMarkAddFirst(schema.marks.underline),
+    'Mod-U': toggleMarkAddFirst(schema.marks.underline),
+    Enter: chainCommands(
+      newlineInCode,
+      createParagraphNear,
+      liftEmptyBlock,
+      splitBlock,
+      insertHardBreak,
+    ),
   };
 };